seeder for achievements

+++ app/database/migrations/2015_09_22_185335_create_challenges_table.php
@@ -36,7 +36,9 @@ class CreateChallengesTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('challenges');
+		if (Schema::hasTable('challenges')) {
+			Schema::drop('challenges');
+		}
 	}
 
 }
