+++ app/database/migrations/2016_03_26_071707_create_achievements_table.php
@@ -33,7 +33,9 @@ class CreateAchievementsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('achievements');
+		if (Schema::hasTable('achievements')) {
+			Schema::drop('achievements');
+		}
 	}
 
 }
