+++ app/database/migrations/2016_03_26_095426_create_user_achievements_mm_table.php
@@ -35,7 +35,9 @@ class CreateUserAchievementsMmTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('users_achievements_mm');
+		if (Schema::hasTable('users_achievements_mm')) {
+			Schema::drop('users_achievements_mm');
+		}
 	}
 
 }
